Move files around to be conformant with the java example.
