--- src/python/twitter/pingpong/__init__.py
@@ -1,5 +1,0 @@
-__all__ = (
-    'PingPongServer'
-)
-
-from .pingpong import PingPongServer
